json web token added

+++ controllers/contacts_controller.js
@@ -4,6 +4,9 @@ import mongoose from "mongoose";
 
 export const getContacts = async (req, res) => {
   try {
+    if (req.role === !process.env.ADMIN) {
+      return res.status(401).json({ message: "UNAUTHORIZED" });
+    }
     const contacts = await Contact.find();
     res.status(200).json(contacts);
   } catch (error) {
@@ -15,6 +18,9 @@ export const createContact = async (req, res) => {
   const contact = req.body;
   const newContact = new Contact(contact);
   try {
+    if (!req.userId) {
+      return res.status(401).json({ message: "unauthenticated" });
+    }
     await newContact.save();
     res.status(201).json(newContact);
   } catch (err) {
@@ -27,7 +33,11 @@ export const updateContact = async (req, res) => {
   const contact = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(_id)) {
-    return res.status(404).send("No post with that id");
+    return res.status(404).send("No contact with that id");
+  }
+
+  if (req.role === !process.env.ADMIN) {
+    return res.status(401).json({ message: "UNAUTHORIZED" });
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -45,7 +55,11 @@ export const deleteContact = async (req, res) => {
   const { id } = req.params;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res.status(404).send("No post with that id");
+    return res.status(404).send("No contact with that id");
+  }
+
+  if (req.role === !process.env.ADMIN) {
+    return res.status(401).json({ message: "UNAUTHORIZED" });
   }
 
   await Contact.findByIdAndRemove(id);
